backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trading-agent-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "websockets",
    "requests",
    "apscheduler",
    "pandas>=2.3.3",
    "ccxt>=4.0.0",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["main.py"]
